+++ CHANGELOG.md
@@ -18,3 +18,13 @@ Initial Release
 ## [v0.0.3]
 
 - Update readme and gif
+
+## [v1.0.0]
+
+- Change the parser to highlight based on the current block instead of the endOf line character
+- Add support for multiple languages using language specific configuration
+
+## [v1.0.1]
+
+- Updated the readme to include the new features
+- Fixed a bug where the extension would not follow the background color set by the user (#1).